// api/model/userPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * AI Shop Helper
 * OpenAPI spec version: 0.1.0
 */

/**
 * Schema for returning user data to the client.
 */
export interface UserPublic {
  id: string;
  email: string;
  name: string;
  is_active: boolean;
  is_superuser: boolean;
  created_at: string;
  updated_at: string;
}